Defer linkage.
Fixes #3.

Drops all intralibrary linkage and allow the libraries to contain unresolved symbols. This should be fine as long as linking using pkgconfig as it will pull the required libraries automatically but if linking manually using "-lcap_utils-07" it might fail due to unresolved symbols. In that case the user must manually link against the proper libraries or better, switch to using `$(pkg-config --libs libcap_utils-0.7)`.