capdump: store suffix and remove limitation
Within the same run the suffix will be recorded so the lookup will be fast and doesn't need a limitation. If it overrides previous files it might take some time to find the first available it but it probably won't be an issue (e.g. large number of files with the same pattern)